card: adapt the cards to the new design system
They have changed drastically from 1.0 to 1.5, which means we've had
to make a couple changes:

- they're slightly darker grey ;
- they have a border because it's nicer ;
- they use some custom padding and colour for the image because the
  white+grey variant we used previously is gone from the design
  system.

Also in this commit:

Previously we did a bunch of tricks to figure out the current
startups (`author.startups`) and the previous
one (`author.previously`) and then concat the two and show an extra
label if there were more than 5, and show all previous ones with an
inactive tag.

I take it on me this is absolutely non-essential and does not justify
the complexity of the previous code.